+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Change endpoint to request instances UIIDs for report generation. UIDEXP-28.
 * Migrate to `stripes` `v3.0.0` and move `react-intl` to peerDependencies.
 * Add new options with icons to action menu - `Export instances (MARC)` and `Export instance (JSON)`. UIDEXP-33.
+* Add support for unconnected preceding titles. Refs UIIN-950.
 
 ## [1.13.1](https://github.com/folio-org/ui-inventory/tree/v1.13.1) (2019-12-11)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.0...v1.13.1)